Extract GrapplingHook - first pass. Still needs work!
